Aggregate .tsv file results by page title in the .json output
The input .tsv files may contain multiple image suggestions for
a single page. Any such suggestions will be stored in adjacent
rows in the .tsv. Combine any such suggestions into one entry
in the corresponding .json file, so they may be conveniently
processed and returned to callers as a group.

Bug: T272721
Change-Id: Ia53acca4ce067a7e0e9220c8ad7fd2d063b973ce