Sqlite version change
Toolforge and WMF docker images have node version 10.x available,
and the previous sqlite npm package being used was only able to be
used on node versions 11.x and above. This patch uses a different
sqlite package, and in doing so resulted in changing the async
function calls to use async/await as opposed to .then() syntax.

Bug: T276863
Change-Id: I5d2c4833297e3a2cf8833cbde7e8e4780853b025