Filter out pages with no suggestions (all source types)
Some pages will not have any image suggestions. Filter these
out of the response, possibly leaving the response with fewer
than expected pages, or possibly none at all. It would be
preferable to always find the expected number of pages with
suggestions, but that is left for a future change.

Bug: T276993
Change-Id: I231b000f9c750aa49402d93b6a97b481919d6320