Init row_num counts from db tables
Previously we were initializing row_num counts for the
deterministic randomization using the tsv files in /static
as the source of truth for that. We should prefer what is actually in the
database file instead, as thats the actual source of truth.
Without this change, hitting an endpoint for a language that
didn't have a .tsv in the /static directory (but had a table in the
database) would cause the API to crash.

Change-Id: Ia645938a1299e27df5c9e60e3bd1a72379871301