Fix a bug where the offset parameter was being applied twice
The offset parameter was considered when determining the set of
pseudorandom pages to return, and also in the database query.
This resulted in offsetting twice, which for certain query
parameter combinations could incorrectly cause no results to
be returned. Changed to only offset in the query if seed=0 is
specified (in that case, the database should apply the offset
as we're skipping randomization). Also added a noFilter query
parameter to help debug similar issues in the future.

Bug: T276993
Change-Id: I20a266f75f5992e042a91ee6d10c1a70a7e36306