Allow limit to be less than table row count
Previously, we were always using the number of rows in a given
table as a factor to get random row numbers. However if
the number of table rows < limit given in the
request, getPsuedoRandomNums would never fill the requested
number of random rows. If the number of table rows < limit,
use limit as the factor of multiplication.

Bug: T286469
Change-Id: I25771496ad7d705ccd11ab64d062b9af36582327