Handle duplicate images
Previously, image filename was being used as a PK for the image table.
However, there may be multiple images with the same filename but with
different sources/confidence levels, and therefore cant be used as the
primary key.

Additionally, we may encounter a duplicate image and page combination
(potential dirty data). In that case, log the occurence, and retry with a new
insertOrIgnore() function which will ignore any violated unique
constraints.

Bug: T285693
Change-Id: I3d7c53231efa4ec02c6721436fc3bfc3cb5faac5